+++ packages/react/src/index.ts
@@ -152,6 +152,9 @@ export { default as Tree } from './components/Tree'
 
 export { default as Indicator, IndicatorProps } from './components/Indicator/Indicator'
 
+export { default as Reaction, ReactionProps } from './components/Reaction/Reaction'
+export { default as ReactionGroup, ReactionGroupProps } from './components/Reaction/ReactionGroup'
+
 //
 // Accessibility
 //